[torch] Add folder for torch.aten.*.Scalar comparisons
This folds small version of the tensor-scalar comparison operators as
they are commonly used for shape computations. This includes le, lt, ge,
gt, eq, and ne.

+++ projects/pt1/python/torch_mlir_e2e_test/test_suite/elementwise.py
@@ -413,7 +413,7 @@ class ElementwiseWhereScalarModule(torch.nn.Module):
         ([-1, -1, -1], torch.float32, True),
     ])
     def forward(self, a):
-        return torch.where(a > 0.5, 4.0, 8.0)
+        return torch.where(a > 0.5, 4.0, 8.0).to(torch.float)
 
 
 @register_test_case(module_factory=lambda: ElementwiseWhereScalarModule())
